Update the comments to reflect new bias correction
The sequence-specific bias correction approach is no longer based on the
approach taken in Kallisto, and is, instead, much closer to the approach
of the updated Cufflinks (as described in Roberts et al. 2011).  This
model is also very similar to that used in eXpress.  In a future commit
we should remove the remnants of the Kallisto-esque model.