Fix behavior of --incompatPrior 0
This commit fixes the behavior when incompatPrior is set to 0.
Previously, there was strange behavior (and the behavior was different
between alignment and read based mode), because of a double-precision
floating point comparison close to machine precision.  This implements
special (and intuitive) behavior when the incompat prior is 0 ---
incompatible mappings are ignored completely.  This commit addresses