Allow unqualified specification of mapping path
This commit fixes a bug that would require the file
where the mapping information is to be written to be
qualified (i.e. it must start with a directory or
"./" if the file was to be written in the current
directory).  Now this restriction is lifted, and the
path is converted to absolute form before any other
processing is done.