Fix nan that caused intermittent hanging in Gibbs sampler
This should *hopefully* resolve #111:

There was an issue with the initialization of some prior abundance
values in the Gibbs sampler, under the VB optimization mode.  This
was causing (in issue 111), values of nan to be fed to
std::gamma_distribution().  This, in turn, caused gamma_distribution
to hang indefinitely when compiled with `-Ofast`, due to the
-ffinite-math-only flag.  The prior initalization should be fixed
now so that nan never occurs here.  For the time being, I'm also
changing the compiler flag to -O3 rather than -Ofast, since
the strange math behavior is probably not worth any marginal speed
improvements.  If we want to re-enable this in the future, we can
disable the finite-math-only with -fnofinite-math-only.