Alter behavior when too few reads are assigned
As @schelhorn mentions, when too few reads are assigned,
this causes Salmon to exit with status 1.  A more graceful
solution is to print a warning, output a quant file with
0 TPM / reads for all targets, and to record this failure
someway.

This commit does this by raising an exception if too few
reads ( < 10 by default) are assigned to transcripts in the
index.  In this case, a quant file is populated with 0s for
the TPM and NumReads field for all transcripts.  Also, a warning
is printed to the console and log.  Finally, in the meta_info.json
file, a new field "quant_errors" has been added.  If this
error was triggered, this list will contain the string
"insufficient_assigned_fragments"; otherwise it will be
empty (other errors can be added in the future via this
mechanism).  This commit addresses issue #130.