New implementation when --gcSizeSamp > 1
I've replaced the prior implementation, which sampled the
GC content at the chosen interval, and counted the nucleotides
at the intermediate locations with one that instead uses a
bitvector and fast rank operations to determine the GC count
in a given range.  Some benchmarks show that the rank-based
implementation is both faster than the sampling-based
implementation (even when the sampling size is 2), and uses less
memory.  When --gcSizeSamp is 1 (the default), the previous
dense implemntation remains (and is a bit faster), but for any
--gcSizeSamp greater than 1, the rank-based implementation is
used.  In the future, we'll consider replacing --gcSizeSamp
with a more descriptive flag --frugalGCCounter or some such.