Bump oneTBB v. fix bug in #514
This bumps the oneTBB version that is downloaded to
2020_U2, and sets the min required version to 2019
(actually, we need >= 2019_U4, but it's hard to check that).

This also fixes the bug mentoined in #514, where
using `--posBias` alone (in selective alignment mode)
would lead to a segfault.  It turns out we have an optimization
to not load the ascii representation of transcripts into the
transcript objects (just keep it in the index) if the user
requests no bias correction or _only_ positional bias.  However,
the code that performs the bias correction could then attempt
to reverse complement the non-existent sequence.  When
either --seqBias or --gcBias were enabled with --posBias,
the sequence was loaded and so this bug wasn't triggered.
Now, we added an explicit check so that we don't attempt to
touch sequence that we don't need (and that doesn't exist), if
we are in using only positional bias.  Thanks @jasonvrogers
for reporting this one!