replace deprecated tbb functionality.
This commit removes references to the
tbb/atomic header and tbb/task_scheduler_init
header.  The replacement for tbb::atomic
is std::atomic, which is trivial for
built-in types, but required some work
for doubles.

@k3yavi & @hiraksarkar : please look
over this commit specifically for replacements
of tbb::atomic with std::atomic as it
relates to alevin (SalmonAlevin,
CollapsedCellOptimizer, etc.), but also
please look over in general.  The biggest
change / pita is how std::vector<std::atomic<T>>
requires nothing that uses the copy constructor.