Fix possible NaN in marginal seq bias terms
This adds a small prior to the marginal and
probability matrices in the sequence-specific bias
model to avoid the problem of having a NaN appear
as a marginal probability if there were no observations
with that nucleotide in the given position.  Note that
there are no NaNs in the underlying probability matrix
as used in the VLMM itself (since we check if a
probability is 0 and set to a small log prob if so).
However, the computation of the marginal was done
before this transformation and could be NaN of a nucelotide
was not at all present in a given position.  Thanks
to @PRopp42 for finding and reporting this issue.