guard: check record type
If the user passed the `--writeQualities` flag and we are writing
the mappings to an output file, check to make sure the records actually
contain quality values (i.e. the input is not in FASTA format). If
they do not, then print a warning and disable writing of the quality
scores.